move `ApplicationBuilder` to its own file
