+++ js/packages/teams-ai/src/Application.spec.ts
@@ -2,7 +2,6 @@ import { strict as assert } from 'assert';
 import { Activity, ActivityTypes, Channels, CloudAdapter, MemoryStorage, MessageReactionTypes, TestAdapter } from 'botbuilder';
 import {
     Application,
-    ApplicationBuilder,
     ConversationUpdateEvents,
     MessageReactionEvents,
     TeamsMessageEvents
@@ -21,14 +20,6 @@ describe('Application', () => {
         planner: new TestPlanner()
     };
     const botAppId = 'testBot';
-    const authenticationSettings = {
-        settings: {
-            testSetting: {
-                connectionName: 'testConnectionName',
-                title: 'testTitle'
-            }
-        }
-    };
     const longRunningMessages = true;
     const removeRecipientMention = false;
     const startTypingTimer = false;
@@ -73,53 +64,6 @@ describe('Application', () => {
             assert.equal(app.options.startTypingTimer, startTypingTimer);
             assert.equal(app.options.storage, storage);
             assert.deepEqual(app.options.taskModules, taskModules);
-        });
-    });
-
-    describe('applicationBuilder', () => {
-        it('should create an Application with default options', () => {
-            const app = new ApplicationBuilder().build();
-            assert.notEqual(app.options, undefined);
-            assert.equal(app.options.adapter, undefined);
-            assert.equal(app.options.botAppId, undefined);
-            assert.equal(app.options.storage, undefined);
-            assert.equal(app.options.ai, undefined);
-            assert.equal(app.options.authentication, undefined);
-            assert.equal(app.options.adaptiveCards, undefined);
-            assert.equal(app.options.taskModules, undefined);
-            assert.equal(app.options.removeRecipientMention, true);
-            assert.equal(app.options.startTypingTimer, true);
-            assert.equal(app.options.longRunningMessages, false);
-        });
-
-        it('should create an Application with custom options', () => {
-            const app = new ApplicationBuilder()
-                .setRemoveRecipientMention(removeRecipientMention)
-                .withStorage(storage)
-                .withAIOptions(ai)
-                .withLongRunningMessages(adapter, botAppId)
-                .withAdaptiveCardOptions(adaptiveCards)
-                .withAuthentication(adapter, authenticationSettings)
-                .withTaskModuleOptions(taskModules)
-                .setStartTypingTimer(startTypingTimer)
-                .build();
-            assert.notEqual(app.options, undefined);
-            assert.equal(app.options.adapter, adapter);
-            assert.equal(app.options.botAppId, botAppId);
-            assert.equal(app.options.storage, storage);
-            assert.equal(app.options.ai, ai);
-            assert.equal(app.options.adaptiveCards, adaptiveCards);
-            assert.equal(app.options.authentication, authenticationSettings);
-            assert.equal(app.options.taskModules, taskModules);
-            assert.equal(app.options.removeRecipientMention, removeRecipientMention);
-            assert.equal(app.options.startTypingTimer, startTypingTimer);
-            assert.equal(app.options.longRunningMessages, longRunningMessages);
-        });
-
-        it('should throw an exception if botId is an empty string for longRunningMessages', () => {
-            assert.throws(() => {
-                new ApplicationBuilder().withLongRunningMessages(adapter, '').build();
-            });
         });
     });
